main.py

from airtable import Airtable
import requests
import os
import time

# Load API keys and IDs from environment variables
AIRTABLE_API_KEY = os.getenv('AIRTABLE_API_KEY')
BASE_ID = os.getenv('BASE_ID')
TABLE_NAME = os.getenv('TABLE_NAME')

# Initialize Airtable
airtable = Airtable(BASE_ID, TABLE_NAME, api_key=AIRTABLE_API_KEY)

def fetch_keywords():
    """ Fetch keywords from Airtable """
    records = airtable.get_all()
    return [record['fields']['Keyword'] for record in records if 'Keyword' in record['fields']]

def fetch_serp_results(keyword):
    """ Fetch SERP results for a keyword (placeholder function) """
    # Add your logic to fetch SERP results here
    print(f"Fetching results for: {keyword}")
    time.sleep(1)  # Delay to respect rate limits
    return []

def main():
    keywords = fetch_keywords()
    for keyword in keywords:
        results = fetch_serp_results(keyword)
        # Process and/or store results

if __name__ == "__main__":
    main()
