requirements.txt

airtable-python-wrapper
requests
python-dotenv
